Change class hierarchy
