+++ src/vfx/index.ts
@@ -1,0 +1,6 @@
+//
+// ASA Visual Effects
+//
+export {Effect, EffectParameterObject, EffectEmitterParameterObject, EffectValue, EmitterOperation, createEffect} from "./Effect";
+export {Emitter, EmitterParameterObject, ParticleInitialParameterObject} from "./Emitter";
+export {Particle, ParticleParameterObject} from "./Particle";
